chore: separate list streams
